Killed warnings

+++ TileMapEditor/Source/Interactables/BoxyBox.cpp
@@ -18,12 +18,12 @@ void BoxyBox::setColor(sf::Color color)
     rect.setFillColor(color);
 }
 
-bool BoxyBox::contains(float x, float y) const
+bool BoxyBox::contains(int x, int y) const
 {
-    if (x >= rect.getPosition().x &&
-        y >= rect.getPosition().y &&
-        x < rect.getPosition().x + rect.getSize().x &&
-        y < rect.getPosition().y + rect.getSize().y)
+    if ((float)x >= rect.getPosition().x &&
+        (float)y >= rect.getPosition().y &&
+        (float)x < rect.getPosition().x + rect.getSize().x &&
+        (float)y < rect.getPosition().y + rect.getSize().y)
         return true;
 
     return false;
